Additional renaming and cleanup; got rid of isEmpty: methods.
git-svn-id: http://dysart.cs.byu.edu/chsvn/trunk@6 a75b6edf-6070-4da8-89c8-9992cc802e51

+++ source/AbstractTree.h
@@ -34,7 +34,7 @@
 @interface AbstractTree : NSObject <Tree>
 {
 	/** A count of how many elements are currently in the tree. */
-	unsigned long count;
+	unsigned int count;
 }
 
 /**
